+++ src/components/PaginationBar/index.tsx
@@ -2,25 +2,15 @@ import { PaginationButton } from './PaginationButton';
 
 interface PaginationProps {
   activePage: number;
-  totalPages: number;
+  rangeMax: number;
+  rangeMin: number;
+  pageNumbers: number[];
   onPage: (pageNumber: number) => void;
   onPrev: () => void;
   onNext: () => void;
 }
 
-export function PaginationBar({ activePage, totalPages, onPage, onPrev, onNext }: PaginationProps) {
-  const getScopedPageNumbers = () => {
-    const rangeMax = Math.min(totalPages, Math.max(1, activePage - 2) + 4);
-    const pageNumbers: number[] = [];
-    const rangeMin = Math.max(1, rangeMax - 4);
-    for (let i = Math.max(1, rangeMax - 4); i <= rangeMax; i++) {
-      pageNumbers.push(i);
-    }
-    return { pageNumbers, rangeMax, rangeMin };
-  };
-
-  const { pageNumbers, rangeMax, rangeMin } = getScopedPageNumbers();
-
+export function PaginationBar({ activePage, pageNumbers, rangeMin, rangeMax, onPage, onPrev, onNext }: PaginationProps) {
   return (
     <div className='flex flex-row py-4'>
       {activePage > rangeMin && <PaginationButton value='<' onClick={onPrev} />}