+++ src/context/index.tsx
@@ -1,1 +1,4 @@
+export { KeyBindingContext, KeyBindingContextProvider } from './KeyBinding';
 export { TodoContext, TodoContextProvider } from './Todo';
+
+export type { KeyBindingObject } from './KeyBinding';