+++ src/db/utils/db.ts
@@ -74,11 +74,7 @@ export class Db {
       .limit(itemsPerPage)
       .offset(offset);
 
-    const count = await baseCountQuery
-      .where(where)
-      .orderBy(...orderBy)
-      .limit(itemsPerPage)
-      .offset(offset);
+    const count = await baseCountQuery.where(where);
 
     return { count: count[0].count, result };
   }
@@ -88,7 +84,7 @@ export class Db {
   }
 
   public static async deleteJournal(recordId: Journal['id']) {
-    return await DbBase.client.update(journal).set(sd<undefined>()).where(eq(journal.id, recordId));
+    return await DbBase.client.update(journal).set(sd(undefined, true)).where(eq(journal.id, recordId));
   }
 
   public static async updateJournal(recordId: Journal['id'], record: NewJournal) {
@@ -148,11 +144,7 @@ export class Db {
       .limit(itemsPerPage)
       .offset(offset);
 
-    const count = await baseCountQuery
-      .where(where)
-      .orderBy(...orderBy)
-      .limit(itemsPerPage)
-      .offset(offset);
+    const count = await baseCountQuery.where(where);
 
     return { count: count[0].count, result };
   }
@@ -202,11 +194,7 @@ export class Db {
       .limit(itemsPerPage)
       .offset(offset);
 
-    const count = await baseCountQuery
-      .where(where)
-      .orderBy(...orderBy)
-      .limit(itemsPerPage)
-      .offset(offset);
+    const count = await baseCountQuery.where(where);
 
     return { count: count[0].count, result };
   }